requirements.txt
# Core Framework
streamlit

# FastAPI Backend
fastapi
uvicorn[standard]
websockets

# LLM & RAG Stack
langchain
langchain-google-genai
google-generativeai

# Vector Database
chromadb

# PDF Processing
pypdf

# Visualization
plotly

# Environment Management
python-dotenv

# Utilities
numpy
pandas
